refactor(plugins): move errorPlugin to src/plugins/errors/ (CP-1)
- src/lib/errors/error-plugin.ts -> src/plugins/errors/error-plugin.ts
- Tests moved alongside. lib/errors/ keeps only error classes
  (base-error.ts, http-errors.ts, subscription-errors.ts,
  employee-status-errors.ts) — moving those to modules is CP-5.
- Relative import "./base-error" rewritten to "@/lib/errors/base-error".
- 6 import sites updated (index.ts, test/support, test/helpers,
  ratelimit test, logger-plugin test, moved error-plugin test).

Tests: 45 pass across plugins/errors, plugins/logger, plugins/health,
and lib/ratelimit integration tests.
