+++ src/index.ts
@@ -10,7 +10,6 @@ import { env, isProduction } from "./env";
 import { betterAuthPlugin, OpenAPI } from "./lib/auth-plugin";
 import { parseOrigins } from "./lib/cors";
 import { cronPlugin } from "./lib/cron-plugin";
-import { errorPlugin } from "./lib/errors/error-plugin";
 import { logger } from "./lib/logger";
 import { setupGracefulShutdown } from "./lib/shutdown/shutdown";
 import { adminController } from "./modules/admin";
@@ -22,6 +21,7 @@ import { organizationController } from "./modules/organizations";
 import { paymentsController } from "./modules/payments";
 import { registerPaymentListeners } from "./modules/payments/hooks/listeners";
 import { publicController } from "./modules/public";
+import { errorPlugin } from "./plugins/errors/error-plugin";
 import { healthPlugin } from "./plugins/health/health-plugin";
 import { loggerPlugin } from "./plugins/logger/logger-plugin";
