+++ src/plugins/logger/CLAUDE.md
@@ -8,7 +8,7 @@ O sistema de logging tem duas camadas com responsabilidades distintas:
 |-----------|---------|-----------------|
 | `loggerPlugin` | `src/plugins/logger/logger-plugin.ts` | Access log (todas as requests), `X-Request-ID` header, `requestId` no contexto |
 | `logger` (Pino util) | `src/lib/logger.ts` | Instância Pino + `shouldIgnore` + `generateRequestId` (utilitário puro, sem lifecycle Elysia) |
-| `errorPlugin` | `src/lib/errors/error-plugin.ts` | Error log (detalhes de 5xx e unhandled), formatação do envelope de erro |
+| `errorPlugin` | `src/plugins/errors/error-plugin.ts` | Error log (detalhes de 5xx e unhandled), formatação do envelope de erro |
 | `request-context` | `src/lib/request-context.ts` | Propagação do `requestId` via `AsyncLocalStorage` para qualquer camada |
 
 ### Fluxo de uma request
